Move GrounData out of the area package
