+++ common/src/main/java/net/theevilreaper/tamias/common/ground/GroundDataRegistry.java
@@ -1,6 +1,5 @@
-package net.theevilreaper.tamias.common.area.ground;
+package net.theevilreaper.tamias.common.ground;
 
-import net.theevilreaper.tamias.common.area.GroundData;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.UnmodifiableView;
